requirements.txt
python-dotenv
requests
flask
jsonify
beautifulsoup4
gunicorn
Jinja2